+++ CHANGELOG.rst
@@ -1,9 +1,9 @@
 =========
 Changelog
 =========
-`Version 0.4.1`
-===============
-* 
+`Unreleased`
+================================================================================================
+* 🧮 Evaluates memory usage for specific nodes from `callback.log` files
 
 `Version 0.4.0: Goodbye Singularity Hub <https://github.com/FCP-INDI/cpac/releases/tag/v0.4.0>`_
 ================================================================================================